Multiplication result converted to larger type
CWE-190
CWE-192
CWE-197
CWE-681

To fix this correctly, force the multiplication to be evaluated in size_t
before any multiplication happens.

Best minimal change: cast one multiplicand to size_t in the memcpy length
expression in CDynStatisticImage::DrawImage().

Concretely in src/webserver/src/WebServer.cpp around line 1462:

- Change memcpy(m_img_data, m_background, m_width*m_height*3);
- To memcpy(m_img_data, m_background, static_cast<size_t>(m_width) * m_height * 3u);

This preserves functionality while preventing intermediate int overflow.

No new methods or dependencies are needed.

+++ src/webserver/src/WebServer.cpp
@@ -1459,7 +1459,8 @@ CDynStatisticImage::~CDynStatisticImage()
 void CDynStatisticImage::DrawImage()
 {
 	// copy background first
-	memcpy(m_img_data, m_background, m_width*m_height*3);
+	memcpy(m_img_data, m_background,
+		static_cast<size_t>(m_width) * m_height * 3u);
 
 	//
 	// Now graph itself
